pyproject.toml
[tool.poetry]
name = "cloud-run-mlops-template"
version = "0.1.0"
description = "A minimal project template for machine learning development, API deployment to Google Cloud Run and monitoring."
authors = ["danieldelc <me@danids.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cloud_run_mlops_template"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.82.0"
google-cloud-bigquery = "^3.3.2"
pydantic = "^1.10.2"
scipy = "^1.9.1"
uvicorn = "^0.18.3"
pandas = "^1.4.4"
scikit-learn = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
requests = "^2.28.1"
datamodel-code-generator = "^0.13.1"
ipykernel = "^6.15.2"
matplotlib = "^3.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
